csrftoken

+++ APP/src/Controller/PinsController.php
@@ -18,7 +18,7 @@ class PinsController extends AbstractController
      */
     public function index(PinRepository $pinRepository): Response
     {
-        $pins = $pinRepository->findBy([],["createdAt"=>"DESC"]);
+        $pins = $pinRepository->findBy([], ["createdAt" => "DESC"]);
         /*   dd($pins); */
         return $this->render('pins/show_pins_front.html.twig', [
             'pins' => $pins,
@@ -43,10 +43,10 @@ class PinsController extends AbstractController
     public function ShowAddPage(Request $request, EntityManagerInterface $em): Response
     {
         $pin = new Pin;
-        $form = $this->createForm(PinType::class,$pin);
+        $form = $this->createForm(PinType::class, $pin);
 
         $form->handleRequest($request);
-        
+
         if ($form->isSubmitted() && $form->isValid()) {
             $em->persist($pin);
             $em->flush();
@@ -59,34 +59,34 @@ class PinsController extends AbstractController
     /**
      * @Route("/pins/{id<[0-9]+>}/edit", name="edit_pin" ,methods={"GET","PUT"})
      */
-    public function EditPin(Pin $pin ,Request $request,EntityManagerInterface $em): Response
+    public function EditPin(Pin $pin, Request $request, EntityManagerInterface $em): Response
     {
-            $form = $this->createForm(PinType::class ,$pin, [
-                'method' => 'PUT'
-            ]);
-            $form->handleRequest($request);
-           
+        $form = $this->createForm(PinType::class, $pin, [
+            'method' => 'PUT'
+        ]);
+        $form->handleRequest($request);
+
         if ($form->isSubmitted() && $form->isValid()) {
             $em->flush();
-            return $this->redirectToRoute("app_pin_show_datails",[ 'id' => $pin->getId()]);
+            return $this->redirectToRoute("app_pin_show_datails", ['id' => $pin->getId()]);
         }
 
         return $this->render('pins/edit_pin.html.twig', [
-            "pin"=>$pin,
+            "pin" => $pin,
             "createForm" => $form->createView(),
         ]);
     }
 
-     /**
+    /**
      * @Route("/pins/{id<[0-9]+>}/delete", name="delete_pin" ,methods={"DELETE"})
      */
-    public function DeletePin(Pin $pin ,Request $request,EntityManagerInterface $em): Response
+    public function DeletePin(Pin $pin, Request $request, EntityManagerInterface $em): Response
     {
-
-        
-         $em->remove($pin);
-         $em->flush();
-         return $this->redirectToRoute("app_page");
-
+        /* dd($request->request->get("token")); */
+        if ($this->isCsrfTokenValid("pin_delete" . $pin->getId(), $request->request->get("token"))) {
+            $em->remove($pin);
+            $em->flush();
+        }
+        return $this->redirectToRoute("app_page");
     }
 }
